+++ examples/next-openai/app/use-chat-streaming-tool-calls/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ToolInvocation } from 'ai';
+import { getToolInvocations, ToolInvocation, UIMessage } from 'ai';
 import { Message, useChat } from '@ai-sdk/react';
 
 export default function Chat() {
@@ -30,28 +30,35 @@ export default function Chat() {
           <div key={m.id} className="whitespace-pre-wrap">
             {isNewRole && <strong>{`${m.role}: `}</strong>}
             {m.content}
-            {m.toolInvocations?.map((toolInvocation: ToolInvocation) => {
-              const { toolCallId, args } = toolInvocation;
+            {/** TODO switch to parts */}
+            {getToolInvocations(m as UIMessage).map(
+              (toolInvocation: ToolInvocation) => {
+                const { toolCallId, args } = toolInvocation;
 
-              // render display weather tool calls:
-              if (toolInvocation.toolName === 'showWeatherInformation') {
-                return (
-                  <div
-                    key={toolCallId}
-                    className="p-4 my-2 text-gray-500 border border-gray-300 rounded"
-                  >
-                    <h4 className="mb-2">{args?.city ?? ''}</h4>
-                    <div className="flex flex-col gap-2">
-                      <div className="flex gap-2">
-                        {args?.weather && <b>{args.weather}</b>}
-                        {args?.temperature && <b>{args.temperature} &deg;C</b>}
+                // render display weather tool calls:
+                if (toolInvocation.toolName === 'showWeatherInformation') {
+                  return (
+                    <div
+                      key={toolCallId}
+                      className="p-4 my-2 text-gray-500 border border-gray-300 rounded"
+                    >
+                      <h4 className="mb-2">{args?.city ?? ''}</h4>
+                      <div className="flex flex-col gap-2">
+                        <div className="flex gap-2">
+                          {args?.weather && <b>{args.weather}</b>}
+                          {args?.temperature && (
+                            <b>{args.temperature} &deg;C</b>
+                          )}
+                        </div>
+                        {args?.typicalWeather && (
+                          <div>{args.typicalWeather}</div>
+                        )}
                       </div>
-                      {args?.typicalWeather && <div>{args.typicalWeather}</div>}
                     </div>
-                  </div>
-                );
-              }
-            })}
+                  );
+                }
+              },
+            )}
           </div>
         );
       })}