+++ packages/ai/core/util/should-resubmit-messages.ts
@@ -1,4 +1,5 @@
 import { UIMessage } from '../types';
+import { getToolInvocations } from '../ui/get-tool-invocations';
 import { extractMaxToolInvocationStep } from './extract-max-tool-invocation-step';
 
 export function shouldResubmitMessages({
@@ -13,17 +14,6 @@ export function shouldResubmitMessages({
   messages: UIMessage[];
 }) {
   const lastMessage = messages[messages.length - 1];
-  console.log(lastMessage);
-
-  console.log(
-    messages.length > originalMessageCount ||
-      extractMaxToolInvocationStep(lastMessage.toolInvocations) !==
-        originalMaxToolInvocationStep,
-    // check that next step is possible:
-    isAssistantMessageWithCompletedToolCalls(lastMessage),
-    // limit the number of automatic steps:
-    (extractMaxToolInvocationStep(lastMessage.toolInvocations) ?? 0) < maxSteps,
-  );
   return (
     // check if the feature is enabled:
     maxSteps > 1 &&
@@ -31,12 +21,13 @@ export function shouldResubmitMessages({
     lastMessage != null &&
     // ensure we actually have new steps (to prevent infinite loops in case of errors):
     (messages.length > originalMessageCount ||
-      extractMaxToolInvocationStep(lastMessage.toolInvocations) !==
+      extractMaxToolInvocationStep(getToolInvocations(lastMessage)) !==
         originalMaxToolInvocationStep) &&
     // check that next step is possible:
     isAssistantMessageWithCompletedToolCalls(lastMessage) &&
     // limit the number of automatic steps:
-    (extractMaxToolInvocationStep(lastMessage.toolInvocations) ?? 0) < maxSteps
+    (extractMaxToolInvocationStep(getToolInvocations(lastMessage)) ?? 0) <
+      maxSteps
   );
 }