+++ packages/provider-utils/src/zod-schema.ts
@@ -1,4 +1,4 @@
-import { JSONSchema7 } from 'json-schema';
+import { JSONSchema7 } from '@ai-sdk/provider';
 import { z } from 'zod';
 import zodToJsonSchema from 'zod-to-json-schema';
 import { jsonSchema, Schema } from './schema';