+++ packages/assemblyai/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/assemblyai
 
+## 0.0.1-canary.1
+
+### Patch Changes
+
+- Updated dependencies [7b3ae3f]
+- Updated dependencies [0ff02bb]
+  - @ai-sdk/provider@2.0.0-canary.12
+  - @ai-sdk/provider-utils@3.0.0-canary.13
+
 ## 0.0.1-canary.0
 
 ### Patch Changes